+++ navigation/src/androidUnitTest/kotlin/com/freeletics/khonshu/navigation/NavigationResultRequestTest.kt
@@ -2,8 +2,8 @@ package com.freeletics.khonshu.navigation
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
+import com.freeletics.khonshu.navigation.internal.StackEntry
 import com.freeletics.khonshu.navigation.test.SimpleRoute
-import com.freeletics.khonshu.navigation.test.simpleRootDestination
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
@@ -12,7 +12,7 @@ internal class NavigationResultRequestTest {
     @Test
     fun `NavigationResultRequest emits events`(): Unit = runBlocking {
         val handle = SavedStateHandle()
-        val owner = NavigationResultRequest(NavigationResultRequest.Key(simpleRootDestination.id, "key"), handle)
+        val owner = NavigationResultRequest(NavigationResultRequest.Key(StackEntry.Id(""), "key"), handle)
 
         owner.results.test {
             handle["key"] = NavigationResult(SimpleRoute(1))
@@ -29,7 +29,7 @@ internal class NavigationResultRequestTest {
     @Test
     fun `NavigationResultRequest emits results that were delivered before collection`(): Unit = runBlocking {
         val handle = SavedStateHandle()
-        val owner = NavigationResultRequest(NavigationResultRequest.Key(simpleRootDestination.id, "key"), handle)
+        val owner = NavigationResultRequest(NavigationResultRequest.Key(StackEntry.Id(""), "key"), handle)
 
         handle["key"] = NavigationResult(SimpleRoute(1))
 
@@ -41,7 +41,7 @@ internal class NavigationResultRequestTest {
     @Test
     fun `NavigationResultRequest emits results were delivered before and during collection`(): Unit = runBlocking {
         val handle = SavedStateHandle()
-        val owner = NavigationResultRequest(NavigationResultRequest.Key(simpleRootDestination.id, "key"), handle)
+        val owner = NavigationResultRequest(NavigationResultRequest.Key(StackEntry.Id(""), "key"), handle)
 
         handle["key"] = NavigationResult(SimpleRoute(1))
         owner.results.test {